Additional ggplot2 themes using M+ fonts.

This is an add-on pacakge for hrbrthemes pacakge. It provides seven ggplot2 themes based on http://mplus-fonts.osdn.jp/about-en.html fonts.

M+ FONTS

The M+ FONTS are a font family under the Free license. You can use, copy, and distribute them, with or without modification, either commercially or noncommercially. The font family provides multilingual glyphs. The fonts provide Kana, over 5,000 Kanji, Basic Latin, Latin-1 Supplement, Latin Extended-A, and IPA Extensions glyphs. Most of the Greek, Cyrillic, Vietnamese, and extended glyphs and symbols are included too. So the fonts are in conformity with ISO-8859-1, 2, 3, 4, 5, 7, 9, 10, 13, 14, 15, 16, Windows-1252, T1, and VISCII encoding.

One-time setup

Before using this theme you need to call import_mplus() which will import the M+ fonts in your extrafont database. Then you will need to install the M+ fonts in the directory mentioned in the output of import_mplus() as per your operating system's way of installing fonts.