ggplot2 theme based on M+ M Type2 fonts.

http://mplus-fonts.osdn.jp/design.html#mplus_m2 fonts are a combination of fixed-fullwidth M+ Type-2 for Japanese and fixed-halfwidth M+ M Type-2 for alphabets.

theme_ipsum_mplus_m2(base_family = mplus.fontfamilies$`2m`,
  plot_title_family = mplus.fontfamilies$`2m-bold`,
  subtitle_family = mplus.fontfamilies$`2m-medium`,
  strip_text_family = mplus.fontfamilies$`2m-regular`,
  caption_family = mplus.fontfamilies$`2m-light`,
  axis_title_family = mplus.fontfamilies$`2m-light`, ...)

Arguments

base_family

base font family and size

plot_title_family

plot title family, face, size and margi

subtitle_family

plot subtitle family, face and size

strip_text_family

facet label font family, face and size

caption_family

plot caption family, face, size and margin

axis_title_family

axis title font family, face and size

...

Arguments passed on to hrbrthemes::theme_ipsum

base_size

base font family and size

plot_title_face

plot title family, face, size and margi

plot_title_size

plot title family, face, size and margi

plot_title_margin

plot title family, face, size and margi

subtitle_family

plot subtitle family, face and size

subtitle_face

plot subtitle family, face and size

subtitle_size

plot subtitle family, face and size

subtitle_margin

plot subtitle margin bottom (single numeric value)

strip_text_face

facet label font family, face and size

strip_text_size

facet label font family, face and size

caption_face

plot caption family, face, size and margin

caption_size

plot caption family, face, size and margin

caption_margin

plot caption family, face, size and margin

axis_title_face

axis title font family, face and size

axis_title_size

axis title font family, face and size

axis_title_just

axis title font justification, one of [blmcrt]

plot_margin

plot margin (specify with ggplot2::margin)

grid

panel grid (TRUE, FALSE, or a combination of X, x, Y, y)

axis

add x or y axes? TRUE, FALSE, "xy"

ticks

ticks if TRUE add ticks

Why Arial Narrow?

First and foremost, Arial Narrow is generally installed by default or readily available on any modern system, so it's "free"-ish; plus, it is a condensed font with solid default kerning pairs and geometric numbers.

Building upon <code>theme_ipsum</code>

The function is setup in such a way that you can customize your own one by just wrapping the call and changing the parameters. See source for examples.

Gotchas

There are distinctions between font names and various devices. Names that work for display graphics devices and bitmap ones such as png may not work well for PostScript or PDF ones. You may need two versions of a font-based theme function for them to work in a particular situation. This situation usually only arises when using a newer font with many weights but somewhat irregular internal font name patterns.